+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'staringatlights:infinite-scroll',
-  version: '0.5.0',
+  version: '0.6.0',
   // Brief, one-line summary of the package.
   summary: 'Template level infinite scrolling.',
   // URL to the Git repository containing the source code for this package.
@@ -12,7 +12,7 @@ Package.describe({
 
 Package.onUse(function(api) {
   api.versionsFrom('1.2');
-  api.use(['jquery', 'less', 'reactive-var', 'templating'], 'client');
+  api.use(['jquery', 'less', 'reactive-var', 'templating', 'tmeasday:publish-counts'], 'client');
   api.addFiles('infiniteScroll.html', 'client');
   api.addFiles('infinite-scroll.js', 'client');
   api.addFiles('infinite-scroll.less', 'client');